api: simplify file structure
