+++ api/deploy_test.go
@@ -3,6 +3,7 @@ package api
 import (
 	"fmt"
 	"io/ioutil"
+	"os"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -164,4 +165,23 @@ func TestDeployServiceFromURL(t *testing.T) {
 	}, <-statuses)
 
 	wg.Wait()
+}
+
+func TestCreateTempFolder(t *testing.T) {
+	a, _ := newAPIAndDockerTest(t)
+	deployer := newServiceDeployer(a)
+
+	path, err := deployer.createTempDir()
+	defer os.RemoveAll(path)
+	require.Nil(t, err)
+	require.NotEqual(t, "", path)
+}
+
+func TestRemoveTempFolder(t *testing.T) {
+	a, _ := newAPIAndDockerTest(t)
+	deployer := newServiceDeployer(a)
+
+	path, _ := deployer.createTempDir()
+	err := os.RemoveAll(path)
+	require.Nil(t, err)
 }
