+++ api/get.go
@@ -1,8 +1,32 @@
 package api
 
-import "github.com/mesg-foundation/core/service"
+import (
+	"github.com/mesg-foundation/core/database/services"
+	"github.com/mesg-foundation/core/service"
+)
 
 // GetService returns service serviceID.
 func (a *API) GetService(serviceID string) (*service.Service, error) {
 	return newServiceGetter(a).Get(serviceID)
+}
+
+// serviceGetter provides functionalities to get a MESG service.
+type serviceGetter struct {
+	api *API
+}
+
+// newServiceGetter creates a new serviceGetter with given api.
+func newServiceGetter(api *API) *serviceGetter {
+	return &serviceGetter{
+		api: api,
+	}
+}
+
+// Get returns service serviceID.
+func (g *serviceGetter) Get(serviceID string) (*service.Service, error) {
+	s, err := services.Get(serviceID)
+	if err != nil {
+		return nil, err
+	}
+	return service.FromService(s, service.ContainerOption(g.api.container))
 }
